fix(calendar): address PR review — edit flow, fail-closed, test split
- Add Edit a rule flow and failure policy to rules management
- Shorten CHANGELOG entry per style guide
- Extract calendar prompt tests to calendar-prompt.test.ts
- Bound management-section test helper at next heading
