+++ CHANGELOG.md
@@ -15,7 +15,7 @@ bus event types) are noted explicitly even in the `0.x` range.
 
 ### Added
 
-- **Calendar scheduling rules** — the calendar specialist loads CEO standing scheduling preferences from `config-store` (`calendar`/`rules`) by key at task start (ceo-inbox parity), with rules-management delegation and prompt-structure tests; standing prefs no longer rely on semantic `memory-query`. (#1223)
+- **Calendar scheduling rules** — calendar agent loads CEO scheduling rules from `config-store` by key at task start, replacing semantic `memory-query`. (#1223)
 - **`progress.resumable`** — typed checkpoint block with bounded accumulator and spill pointer, plus round-trip tests. (#1172)
 - **Document workspace** — `working_documents` / `working_document_links` Postgres store with `WorkingDocsRepo`, OKF frontmatter round-trip, backlink index, and optimistic concurrency (document + per-section). (#1208)
